+++ ntex/CHANGES.md
@@ -4,7 +4,7 @@
 
 * ntex::http: Fix handling of large http messages
 
-* ntex::http: Refine read back-pressure for h1 dispatcher
+* ntex::http: Refine read/write back-pressure for h1 dispatcher
 
 ## [0.1.6] - 2020-04-09